\name{summary}
\alias{summary}
\alias{summary.Explore.WS.Corr}
\alias{summary.Fract.Poly}
\alias{summary.WS.Corr.Mixed}
\alias{summary.WS.Corr.Mixed.SAS}
\title{Summary}
\description{summary}
\usage{
\method{summary}{Explore.WS.Corr}(object, ..., Object)
\method{summary}{Fract.Poly}(object, ..., Object)
\method{summary}{WS.Corr.Mixed}(object, ..., Object)
\method{summary}{WS.Corr.Mixed.SAS}(object, ..., Object)
}
\keyword{internal}
